refactor: Phase B round04 - App CRUD/Config 收敛到 AppRegistryService 单一真相源
